pyproject.toml
[tool.poetry]
name =  "sgs"
version = "0.1.0"
description = "Smart Grocery Search is a demo project that combines vector search, knowledge graphs, and DSPy to power smarter, attribute-aware product discovery in grocery retail."
authors = ["Allan Butler <butler.allan@heb.com>" ]
readme = "README.md"
packages = [
    { include = "sgs" },
]
homepage = "https://gitlab.com/butler.allan-heb/smart-grocery-search"
repository = "https://gitlab.com/butler.allan-heb/smart-grocery-search"

[tool.poetry.dependencies]
python = "3.12.2"
python-json-logger = "^2.0.7"
python-dotenv = "^1.1.1"
uvicorn = "^0.35.0"
pandas = "^2.3.2"
networkx = "^3.5"
sentence-transformers = "^5.1.0"
faiss-cpu = "^1.12.0"
fastapi = "^0.116.1"
dspy = "^3.0.3"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.15"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
autoflake8 = "^0.4.1"
flake8 = "^7.3.0"
coverage = "^7.10.6"
mypy = "^1.17.1"
isort = "^6.0.1"
autoflake = "^2.3.1"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[tool.mypy]
python_version = "3.12"
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
warn_return_any = true

[tool.flake8]
max-line-length = 120
exclude = ".git __pycache__ .mypy_cache .pytest_cache .reports notebooks _resource"
extend-ignore = ["E203"]

[tool.pytest.ini_options]
addopts = "--pdbcls=IPython.terminal.debugger:TerminalPdb -p no:warnings"
junit_family = "xunit2"
norecursedirs = ".git __pycache__ .mypy_cache .pytest_cache .reports notebooks _resource"

[tool.coverage.report]
omit = ["*/__init__.py", ".mypy_cache", ".reports", "notebooks"]
show_missing = true

[tool.black]
line-length = 120
target-version = ["py312"]
exclude = '(\.git|\.mypy_cache|\.pytest_cache|\.reports|build|dist|notebooks|_resource)'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "sgs.__main__:main"

